Allow complex selector. Add verbose option. Fix namespace doc.
Fixed the --selector option to quote the label matching expression
and to be consistent with handling of context and namespace.

Fixed "No pods exists that match" error message.

Added --verbose option to list the names of the matched pods.

Corrected the doc re the --namespace option since it doesn't default to
"default", it defaults to whatever the local config is, as it should.
(The code performs an extra 'defaulting' in that an empty namespace
value is treated as 'default', which probably isn't much use.)

Fixed non-tab indent whitespace in --selector handling to use tabs
like the rest of the code.
Added vim modeline to help prevent future inconsistency.